EXSWHTEC-69 - Implement new tests for hipMemRangeGetAttribute and hipMemRangeGetAttributes.
